Add files via upload
Design a simple cash register similar to one found at McDonald’s or Burger King. To do this,
determine a menu of five or six items from the restaurant. Also, include a Total button or a clear
button, or possibly both. Also, include a means for backing out of a mistake without starting over
from zero. Display the cost of the total order in the PLC at an address in the data table. Use
floating-point math and you are encouraged to do so.